requirements.txt
PyQt5==5.11.3
PyYAML==4.2b4
sip==4.19.8